(fix + design): Unify output-eltype via zero-slope arithmetic (#144)
Treat Constant as `y[idx] + 0·(xq - x[idx])` so `Tq`'s carrier (Dual,
Measurement, SVector, …) propagates naturally — the same rule Linear /
Cubic / Quadratic / Hermite already follow. Drop the scalar
`convert(_output_eltype, val)` wrapper and replace the persistent
allocating batch + 1D Constant oneshot allocators with sample-first so
the container fits whatever the kernel actually produces. Constant
`_output_eltype` overrides and inline `T_out` expressions removed; ND
Constant scalar oneshot drops its `::Tv` typeassert. Every method now
uses the generic `_output_eltype(Tv, Tg, Tq)`.

Constant kernels multiply the selection by `one(dL)` (carrier
propagation); derivative branches use `0 * y_left * one(dL)` to
preserve NaN/Inf propagation and keep the minimal `*(::Int, ::Tv)` +
`*(::Tv, ::Real)` duck-type contract.

Behavior change: `constant_interp` on `Int y` with a `Float xq` now
returns `Float` (carrier propagates through zero-slope arithmetic).
The fully-Int chain `(Int x, Int y, Int xq)` still returns `Int`. The
52a1f91 raw-Tv-for-promotable-Tq contract from #138 is intentionally
relaxed to keep Constant on the same machinery as every other method —
no Constant-specific helpers. Series / ND-oneshot-3-arg / hetero
allocators remain on the old path and will migrate in a follow-up.

New `test_duck_tv_dual_tq.jl` pins the Tv × Tq cross product across
methods. Existing Constant / Complex / precision / promotion tests
updated to the natural-promote contract.

Closes #144.

+++ test/test_precision_vector_queries.jl
@@ -50,10 +50,9 @@
 
         @testset "Allocating vector call" begin
             @test result_vec ≈ result_broadcast rtol = PRECISION_RTOL
-            # Constant is a selection kernel — output keeps raw `Tv` (Float32),
-            # not promoted to the query precision. Scalar / broadcast / batch
-            # all agree on this contract for plain numeric queries.
-            @test eltype(result_vec) == Float32
+            # Zero-slope arithmetic propagates the Float64 xq carrier:
+            # Float32 y * one(Float64) = Float64. Scalar / broadcast / batch agree.
+            @test eltype(result_vec) == Float64
         end
 
         @testset "In-place vector call" begin
